Package refactoring: grails.plugins.jms.test -> grails.plugin.jms.test.
