+++ integration-test-app/grails-app/domain/grails/plugin/jms/test/Person.groovy
@@ -1,0 +1,6 @@
+package grails.plugin.jms.test
+
+class Person {
+    static hasMany = [things: Thing]
+    String name
+}
