+++ integration-test-app/grails-app/domain/grails/plugin/jms/test/Thing.groovy
@@ -1,0 +1,6 @@
+package grails.plugin.jms.test
+
+class Thing {
+    static belongsTo = Person
+    String name
+}
