--- integration-test-app/grails-app/domain/grails/plugins/jms/test/Person.groovy
@@ -1,6 +1,0 @@
-package grails.plugins.jms.test
-
-class Person {
-    static hasMany = [things: Thing]
-    String name
-}
