--- integration-test-app/grails-app/domain/grails/plugins/jms/test/Thing.groovy
@@ -1,6 +1,0 @@
-package grails.plugins.jms.test
-
-class Thing {
-    static belongsTo = Person
-    String name
-}
