--- integration-test-app/grails-app/init/integration/test/app/Application.groovy
@@ -1,14 +1,0 @@
-package jms
-
-import grails.boot.GrailsApp
-import grails.boot.config.GrailsAutoConfiguration
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
-import org.springframework.boot.autoconfigure.jms.JmsAutoConfiguration
-import org.springframework.boot.autoconfigure.jta.JtaAutoConfiguration
-
-@EnableAutoConfiguration(exclude = [JmsAutoConfiguration, JtaAutoConfiguration])
-class Application extends GrailsAutoConfiguration {
-    static void main(String[] args) {
-        GrailsApp.run(Application, args)
-    }
-}
