+++ integration-test-app/grails-app/services/grails/plugin/jms/test/config/OtherListenerService.groovy
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2010 Grails Plugin Collective
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package grails.plugin.jms.test.config
+
+import grails.plugin.jms.test.TestListeningServiceSupport
+
+import javax.jms.Message
+
+class OtherListenerService extends TestListeningServiceSupport {
+
+    static exposes = ['jms']
+    static adapter = "other"
+    static container = "other"
+
+    def onMessage(msg) {
+        putMessage(msg instanceof Message)
+    }
+
+}
