--- integration-test-app/grails-app/services/grails/plugins/jms/test/domain/DomainService.groovy
@@ -1,36 +1,0 @@
-/*
- * Copyright 2010 Grails Plugin Collective
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package grails.plugins.jms.test.domain
-
-import grails.plugins.jms.test.Person
-import grails.plugins.jms.test.TestListeningServiceSupport
-
-class DomainService extends TestListeningServiceSupport {
-
-    static transactional = false
-
-    static exposes = ['jms']
-
-    def onMessage(msg) {
-        try {
-            def person = Person.findByName(msg)
-            putMessage(person.things*.name)
-        }
-        catch (e) {
-            log.error(e.getMessage(), e)
-        }
-    }
-}
