--- integration-test-app/grails-app/services/grails/plugins/jms/test/reply/ReplyingListenerService.groovy
@@ -1,34 +1,0 @@
-/*
- * Copyright 2010 Grails Plugin Collective
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package grails.plugins.jms.test.reply
-
-import grails.plugins.jms.Queue
-import grails.plugins.jms.test.TestListeningServiceSupport
-
-class ReplyingListenerService extends TestListeningServiceSupport {
-
-    static exposes = ['jms']
-
-    @Queue
-    def initial(msg) {
-        1
-    }
-
-    @Queue
-    def reply(msg) {
-        putMessage(msg)
-    }
-}
