+++ integration-test-app/src/integration-test/groovy/grails/plugin/jms/test/domain/DomainServiceSpec.groovy
@@ -1,0 +1,36 @@
+package grails.plugin.jms.test.domain
+
+import grails.plugin.jms.test.Person
+import grails.plugin.jms.test.Thing
+import grails.test.mixin.integration.Integration
+import grails.transaction.Transactional
+import spock.lang.Specification
+
+@Integration
+class DomainServiceSpec extends Specification {
+
+    static transactional = false
+
+    def jmsService
+    def domainService
+
+    @Transactional
+    void setupData() {
+        def p = new Person()
+        p.name = 'p'
+        [1, 2, 3, 4].each {
+            p.addToThings(new Thing(name: it.toString()))
+        }
+        p.save(flush: true, failOnError: true)
+    }
+
+    void testIt() {
+        given:
+        setupData()
+        when:
+        jmsService.send(service: 'domain', 'p')
+        then:
+        domainService.message.sort() == ['1', '2', '3', '4']
+    }
+
+}
