--- integration-test-app/src/integration-test/groovy/grails/plugins/jms/test/servicelevel/ServiceMethodListeningServiceSpec.groovy
@@ -1,18 +1,0 @@
-package grails.plugins.jms.test.servicelevel
-
-import grails.test.mixin.integration.Integration
-import spock.lang.Specification
-
-@Integration
-class ServiceMethodListeningServiceSpec extends Specification {
-
-    def jmsService
-    def serviceMethodListeningService
-
-    void testIt() {
-        when:
-        jmsService.send(service: 'serviceMethodListening', "a")
-        then:
-        serviceMethodListeningService.message == "a"
-    }
-}
